backend/requirements.txt
# backend/requirements.txt (example — update/add other packages you need)
mcp==0.9.1
fastapi==0.116.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
geoalchemy2==0.14.2
postgis==1.0.4

# Langchain + OpenAI adapter (keep langchain version you want)
langchain==0.3.27
langchain-openai==0.3.33

# OPENAI: choose a version that satisfies langchain-openai (>=1.104.2,<2.0.0)
openai==1.104.2

# LangSmith (if you use it)
langsmith==0.4.27

# Common helpers
requests==2.31.0
aiofiles==23.2.1
aiohttp==3.9.1
python-multipart==0.0.6
openpyxl==3.1.2
beautifulsoup4==4.12.2
plotly==5.17.0
matplotlib==3.8.2
pandas==2.1.4
numpy==1.25.2
spacy==3.7.2
python-dotenv==1.0.0
redis==5.0.1

# Ensure pydantic is new enough for langchain, but <3
pydantic>=2.7.4,<3.0.0


